Refactor: Eliminate barrel files in src/ui
Removes index.ts files in src/ui/serve-behaviors and src/ui/copy-behaviors.
Renames src/ui/navigation-behaviors/index.ts to handler.ts.
Updates imports in dependent files (JsonTree.tsx, InteractiveViewer.tsx, handler.test.ts).
Updates PREVENT_CONFLICTS.md to mark the task as completed.
